Former-commit-id: 11bcec66006fee8975c768b9c6fd7d5ac5e4db3f

+++ src/runtime/mini_stdint.h
@@ -1,16 +1,16 @@
 #ifndef MINI_STDINT_H
 #define MINI_STDINT_H
 
-typedef signed char		int8_t;
-typedef short int		int16_t;
-typedef int			int32_t;
-typedef unsigned char		uint8_t;
-typedef unsigned short int	uint16_t;
-typedef unsigned int		uint32_t;
+typedef signed char        int8_t;
+typedef short int          int16_t;
+typedef int                int32_t;
+typedef unsigned char      uint8_t;
+typedef unsigned short int uint16_t;
+typedef unsigned int       uint32_t;
 
 #ifdef BITS_64
-typedef long int		int64_t;
-typedef unsigned long int	uint64_t;
+typedef long long int           int64_t;
+typedef unsigned long long int  uint64_t;
 typedef uint64_t size_t;
 typedef int64_t intptr_t;
 typedef int64_t ptrdiff_t;
@@ -19,8 +19,8 @@ typedef int64_t ptrdiff_t;
 #endif
 
 #ifdef BITS_32
-__extension__ typedef long long int		int64_t;
-__extension__ typedef unsigned long long int	uint64_t;
+__extension__ typedef long long int          int64_t;
+__extension__ typedef unsigned long long int uint64_t;
 typedef uint32_t size_t;
 typedef int32_t intptr_t;
 typedef int32_t ptrdiff_t;
